pyproject.toml
[tool.poetry]
name = "nonebot-plugin-nai3"
version = "0.0.1"
description = "通过 NovelAI 生成图片"
authors = ["zhulinyv <zhulinyv2005@outlook.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "nonebot_plugin_nai3"}]

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.2.1"
httpx = "^0.24.1"
nonebot-adapter-onebot = "^2.2.3"

# https://black.readthedocs.io/en/stable/the_black_code_style/current_style.html
[tool.black]
line-length = 500

# https://beta.ruff.rs/docs/settings/
[tool.ruff]
line-length = 500
# https://beta.ruff.rs/docs/rules/
select = ["E", "W", "F"]
ignore = ["F401"]
# Exclude a variety of commonly ignored directories.
respect-gitignore = true
ignore-init-module-imports = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
